+++ src/zodbupdate/convert.py
@@ -57,18 +57,50 @@ def default_renames():
         ('datetime', 'time'): ('zodbupdate.convert', 'Time')}
 
 
-def decode_attribute(attribute, encoding):
+def decode_attribute(attribute, encoding, encoding_fallbacks=[]):
 
     def decode(data):
         value = data.get(attribute)
-        if value is not None:
-            if isinstance(value, six.text_type):
-                if encoding == utils.ENCODING:
-                    return False
-                value = utils.safe_binary(value)
+        if value is None:
+            return False
+        if isinstance(value, six.text_type):
+            if encoding == utils.ENCODING:
+                return False
+            value = utils.safe_binary(value)
+        try:
             data[attribute] = value.decode(encoding)
-            return True
-        return False
+        except UnicodeDecodeError:
+            for encoding_fallback in encoding_fallbacks:
+                try:
+                    data[attribute] = value.decode(encoding_fallback)
+                except UnicodeDecodeError:
+                    continue
+                logger.warning(
+                    'Encoding fallback to "{fallback_encoding:s}" '
+                    'while decoding attribute "{attribute:s}" '
+                    'from: \n{value}\nto:\n{data_attribute}'.format(
+                        fallback_encoding=encoding_fallback,
+                        attribute=attribute,
+                        value=value,
+                        data_attribute=data[attribute],
+                    )
+                )
+                logger.debug(
+                    "Encoding fallback on data:\n{data}".format(data=data)
+                )
+                return True
+            else:
+                raise UnicodeDecodeError(
+                    'encoding={encoding}, fallback_encodings={fallbacks}'.format(
+                        encoding=encoding,
+                        fallbacks=encoding_fallbacks,
+                    ),
+                    value,
+                    0,
+                    0,
+                    'Neither with encoding nor with fallbacks.',
+                )
+        return True
 
     return decode
 
@@ -85,7 +117,7 @@ def encode_binary(attribute):
     return encode
 
 
-def load_decoders():
+def load_decoders(encoding_fallbacks=[]):
     decoders = {}
     for entry_point in pkg_resources.iter_entry_points('zodbupdate.decode'):
         definition = entry_point.load()
@@ -96,7 +128,7 @@ def load_decoders():
                     encode_binary(attribute))
             else:
                 decoders.setdefault((module, cls), []).append(
-                    decode_attribute(attribute, encoding))
+                    decode_attribute(attribute, encoding, encoding_fallbacks))
         logger.info('Loaded {} decode rules from {}:{}'.format(
             len(definition), entry_point.module_name, entry_point.name))
     return decoders